requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
psycopg2-binary==2.9.9
django-autoslug==1.9.9
drf-yasg==1.21.7
